backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
pydantic==2.5.0
python-multipart==0.0.6
torch==2.1.0
numpy==1.24.3
pandas==2.1.3
ta==0.11.0
python-binance==1.0.19
openai==1.3.5
supabase==2.0.3
cryptography==41.0.7
slowapi==0.1.9
